speed up tests
