+++ smoke-test/tests/knowledge/document_helpers.py
@@ -1,0 +1,16 @@
+import uuid
+
+
+def execute_graphql(auth_session, query: str, variables: dict | None = None) -> dict:
+    """Execute a GraphQL query against the frontend API."""
+    payload = {"query": query, "variables": variables or {}}
+    response = auth_session.post(
+        f"{auth_session.frontend_url()}/api/graphql", json=payload
+    )
+    response.raise_for_status()
+    result = response.json()
+    return result
+
+
+def _unique_id(prefix: str) -> str:
+    return f"{prefix}-{uuid.uuid4().hex[:8]}"
